+++ cv-application/src/components/Header.jsx
@@ -1,4 +1,4 @@
-import './Header.css'
+import '../styles/Header.css'
 
 export default function Header () {
     return (
