+++ cv-application/src/main.jsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import Header from './Header.jsx'
-import App from './App.jsx'
-import Footer from './Footer.jsx'
+import Header from './components/Header.jsx'
+import App from './components/App.jsx'
+import Footer from './components/Footer.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
