+++ cv-application/src/styles/App.css
